+++ kedro_azureml/manager.py
@@ -18,9 +18,9 @@ class KedroContextManager:
         self,
         env: str,
         project_path: Optional[str] = None,
-        extra_params: Optional[dict] = None,
+        runtime_params: Optional[dict] = None,
     ):
-        self.extra_params = extra_params
+        self.runtime_params = runtime_params
         self.env = env
         self.project_path = project_path
         self.session: Optional[KedroSession] = None
@@ -74,11 +74,11 @@ CONFIG_LOADER_ARGS = {
                 raise ValueError(
                     "Missing azureml.yml files in configuration. Make sure that you configure your project first"
                 )
-        return KedroAzureMLConfig.parse_obj(obj)
+        return KedroAzureMLConfig.model_validate(obj)
 
     def __enter__(self):
         self.session = KedroSession.create(
-            self.project_path, env=self.env, extra_params=self.extra_params
+            self.project_path, env=self.env, runtime_params=self.runtime_params
         )
         return self